requirements.txt
aiohttp==3.6.2
appdirs==1.4.3
astroid==2.3.3
async-timeout==3.0.1
asyncpg==0.20.1
attrs==19.3.0
beautifulsoup4==4.9.0
black==19.10b0
cffi==1.14.0
chardet==3.0.4
click==7.1.1
colorama==0.4.3
discord.py==1.3.3
idna==2.9
isort==4.3.21
lazy-object-proxy==1.4.3
lxml==4.5.0
mccabe==0.6.1
multidict==4.7.5
pathspec==0.7.0
Pillow==7.1.1
pycparser==2.20
pylint==2.4.4
PyNaCl==1.3.0
regex==2020.4.4
six==1.14.0
soupsieve==2.0
toml==0.10.0
typed-ast==1.4.1
websockets==8.1
wrapt==1.11.2
yarl==1.4.2
youtube-dl==2020.3.24
